Rewrote JcTreeConverter test to parsing source with (A) javac, and (B) with javac, converting to lombok.ast, and back to javac.
Contrast to old way: parse with (A) lombok, and (B) javac, then convert to lombok.ast.

This way should make it a lot simpler to avoid erroneous failing tests due to e.g. javac folding concatenated string literals together, and losing information such as how a numeric literal is specified (as hex or as decimal, etc).